finalizo la documentacion de toda la carpeta soldados
