+++ Batallon/src/App.java
@@ -1,4 +1,14 @@
- import java.util.ArrayList;
+import java.util.ArrayList;
+import rangos.Nivel_militar;
+//import soldados.Capitan;
+import soldados.Coronel;
+import soldados.Soldado;
+import soldados.SoldadoRaso;
+import soldados.Teniente;
+//import rangos.Nivel_militar;
+// import rangos.Rango;
+// import misiones.OperacionesMilitares;
+
 public class App {
     public static void main(String[] args) {
         // Crear instancias de Soldado, SoldadoRaso, Teniente y Coronel
@@ -11,12 +21,14 @@ public class App {
         //teniente.realizarAccion();
         //coronel.realizarAccion();
 
-        // Prueba de los métodos get y set para cada clase atributos private 
-        //System.out.println("Unidad del teniente: " + teniente.getUnidad());
-        //System.out.println("Estrategia del coronel: " + coronel.getEstrategia());
+        // //Prueba de los métodos get y set para cada clase atributos private 
+        // System.out.println("---------------------------------------");
+        // System.out.println("Unidad del teniente: " + teniente.getUnidad());
+        // System.out.println("Estrategia del coronel: " + coronel.getEstrategia());
         // System.out.println("Nombre del soldado raso: " + soldadoRaso1.getNombre());
-        //System.out.println("ID del soldado raso: " + soldadoRaso1.getId());
-        //System.out.println("Rango del soldado: "+ soldadoRaso1.getRango());
+        // System.out.println("ID del soldado raso: " + soldadoRaso1.getId());
+        // System.out.println("Rango del soldado: "+ soldadoRaso1.getRango());
+        // System.out.println("---------------------------------------");
 
         //Prueba de herencia, metodo mostrarInformacion
         soldadoRaso1.mostrarInformacion();
@@ -24,8 +36,12 @@ public class App {
 
         soldadoRaso1.asignarMision("MISION SALVAR MARTE"); 
         soldadoRaso1.reportarEstado(); 
+        coronel.asignarMision("Armar contraataque enemigo");
+        coronel.reportarEstado();
 
-        System.out.println("el numero de soldados es "+Soldado.getContadorSoldados() );
+
+        System.out.println("---------------------------------------");
+        System.out.println("El numero de soldados es: "+ Soldado.getContadorSoldados() );
 
 
         //Downcasting}
