+++ Batallon/src/rangos/Nivel_militar.java
@@ -1,3 +1,4 @@
+package rangos;
 public enum Nivel_militar {
     SOLDADO, SOLDADO_RASO, TENIENTE, CORONEL 
 }
