+++ Batallon/src/soldados/Soldado.java
@@ -1,11 +1,20 @@
+package soldados;
+import rangos.Nivel_militar;
+import rangos.Rango;
+//import misiones.OperacionesMilitares;
+//imopotamos paquetes necesearios
+
+
+//se implementa la clase base "Soldado" que extiende de la clase abstracta Rango
 public class Soldado extends Rango { 
     
-
+    //Atributos de la clase Soldado
     private final String nombre;
     private final String id;
     protected Nivel_militar rango; // 
     private static int contadorSoldados = 0; // Contador de Soldados
 
+    //Constructor de la clase Soldado que recibe los parametros nivel, nombre, id y rango
     public Soldado(int nivel, String nombre, String id, Nivel_militar rango) {
         super(nivel); 
         this.nombre = nombre;
@@ -13,29 +22,29 @@ public class Soldado extends Rango {
         this.rango = rango;
         contadorSoldados++ ; // Incrementa el contador de Soldados 
     }
-
+    //Metodo que retorna el contador de Soldados, es static para que pueda ser accedido sin instanciar un objeto
     public static int getContadorSoldados() {
         return contadorSoldados;
     } 
 
+    //Metodo que se implementa en las clases hijas
     public void realizarAccion() { 
         // Implementacion de metodo realizar accion
     }
 
+
+    //Metodos Get
     public String getNombre() {
         return nombre;
     }
-
-
     public String getId() {
         return id;
     }
-
-
      public Nivel_militar getRango() {
          return rango;
     }
 
+    //Metodo Set
     public void setRango(Nivel_militar rango) {
         this.rango = rango;
     }
@@ -53,6 +62,9 @@ public class Soldado extends Rango {
         System.out.println("ID: " + id);
         System.out.println("Rango: " + rango);
         System.out.println("Nivel: " + nivel);
+        System.out.println("--------------------------");
+
+    //en el app.java esta el ejemplo de como se imprime la informacion de los soldados
     }
